docs: remove qr code snippets
