+++ apps/compositions/src/examples/qr-code-with-fill.tsx
@@ -1,12 +1,15 @@
-import { Flex, For } from "@chakra-ui/react"
-import { QrCode } from "compositions/ui/qr-code"
+import { Flex, For, QrCode } from "@chakra-ui/react"
 
 export const QrCodeWithFill = () => {
   return (
     <Flex gap="4">
       <For each={["#5417D7", "#FF0000"]}>
         {(fill) => (
-          <QrCode key={fill} fill={fill} value="https://www.google.com" />
+          <QrCode.Root key={fill} value="https://www.google.com">
+            <QrCode.Frame fill={fill}>
+              <QrCode.Pattern />
+            </QrCode.Frame>
+          </QrCode.Root>
         )}
       </For>
     </Flex>
