--- apps/compositions/src/examples/qr-code-without-snippet.tsx
@@ -1,11 +1,0 @@
-import { QrCode } from "@chakra-ui/react"
-
-export const QrCodeWithoutSnippet = () => {
-  return (
-    <QrCode.Root value="..." size="md">
-      <QrCode.Frame>
-        <QrCode.Pattern />
-      </QrCode.Frame>
-    </QrCode.Root>
-  )
-}
